Update README.md
Create a Javascript application :
Below are two list of User
List1=[‘Arjun’, ‘Adwait’, ‘Swapnil’,’Amit’, ‘Vishal’, ‘Adnan’]
List2=[ ‘Adwait’,’Laxman’,’Amit’, ‘Adnan’,’Nitin’,’Gaurav’]
Find following output (Use collection)
a. Get a unique set of users from List1 which are not in List2
b. Get a unique set of users from List2 which are not in List1
c. Get a set of users who are present in in List1 and List2 both (intersection of list1 & list2)